Провел настройку consul для сервисов transfer и cash. Добавил тестов в кеш сервис

+++ cash/src/main/java/com/bank/cash/dto/AccountBalanceDto.java
@@ -2,10 +2,12 @@ package com.bank.cash.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AccountBalanceDto(
         Long id,
         String currency,
-        double balance,
+        BigDecimal balance,
         boolean isExists
 ) {}
