+++ cash/src/main/java/com/bank/cash/dto/AccountBalanceUpdateRequest.java
@@ -2,8 +2,10 @@ package com.bank.cash.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AccountBalanceUpdateRequest(
         String currency,
-        double balance
+        BigDecimal balance
 ) {}
