+++ cash/src/main/java/com/bank/cash/dto/CashRequest.java
@@ -6,9 +6,11 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record CashRequest(
         @NotBlank(message = "Валюта обязательна") String currency,
-        @NotNull(message = "Сумма обязательна") @Positive(message = "Сумма должна быть положительной") Double value,
+        @NotNull(message = "Сумма обязательна") @Positive(message = "Сумма должна быть положительной") BigDecimal value,
         @NotNull(message = "Действие обязательно") CashAction action
 ) {}
